pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "deep-translator>=1.11.4",
    "google-genai>=1.25.0",
    "pillow>=11.3.0",
    "streamlit>=1.46.1",
]
